pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sarlakbot"
version = "3.0.0"
description = "Professional Educational Telegram Bot for Iranian Students | Sarlak Academy"
authors = [
    { name = "Sarlak Academy Team", email = "info@sarlak.academy" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["telegram", "education", "bot", "konkur", "ai", "sarlak", "academy"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Application Frameworks"
]

dependencies = [
    "python-telegram-bot>=21.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "redis>=6.4.0",
    "pydantic>=2.12.3",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "rich>=14.2.0",
    "cryptography>=42.0.5",
    "tenacity>=8.3.0"
]

[project.optional-dependencies]
dev = [
    "black>=25.9.0",
    "ruff>=0.14.1",
    "flake8>=7.1.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=6.0.0",
    "pre-commit>=4.3.0",
    "bandit>=1.8.6",
    "safety>=3.6.2",
    "isort>=5.13.2",
    "types-requests",
    "types-redis",
    "types-python-dateutil"
]

[project.urls]
Homepage = "https://sarlak.academy"
Repository = "https://github.com/alirezasarlak/botsarlak-core"
Documentation = "https://docs.sarlak.academy"
"Bug Tracker" = "https://github.com/alirezasarlak/botsarlak-core/issues"

[tool.black]
line-length = 100
target-version = ['py311']
skip-string-normalization = false

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
    "migrations"
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "ASYNC", "S", "UP", "C4", "Q"]
ignore = ["E501", "E731", "B905"]
fixable = ["ALL"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
